Fix memory clearing in the json proxy cache

+++ nlw/lib/Socialtext/JSON/Proxy/Helper.pm
@@ -6,6 +6,7 @@ use warnings;
 use File::Temp qw/mkdtemp/;
 use Socialtext::Paths;
 use Socialtext::AppConfig;
+use Socialtext::System qw(shell_run);
 use Socialtext::File;
 use File::Find qw(find);
 
@@ -13,11 +14,11 @@ our $cache_dir = Socialtext::Paths::storage_directory('json_cache');
 
 sub ClearMemoryCache {
     my $class = shift;
-    my $pidfile = Socialtext::AppConfig->pid_file_dir . "/json-proxy.pid";
-    if (-f $pidfile) {
-        my $pid = Socialtext::File::get_contents($pidfile);
-        system "kill -USR1 $pid";
-    }
+    my $jsonproxy_svc_dir = Socialtext::AppConfig->is_dev_env
+        ? "$ENV{ST_CURRENT}/plugins/widgets/service"
+        : "/usr/share/nlw/plugin/widgets/service";
+    my ($pid) = `svstat $jsonproxy_svc_dir` =~ /up \(pid (\d+)\)/;
+    system "kill -USR1 $pid" if $pid;
 }
 
 sub ClearForUsers {
